packages/voice/pyproject.toml
[tool.poetry]
name = "zenus-voice"
version = "0.1.0"
description = "Voice interface for Zenus OS - hands-free control with STT and TTS"
authors = ["Zeni <zeni@zenus.ai>"]
readme = "README.md"
packages = [{include = "zenus_voice", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
zenus-core = {path = "../core", develop = true}
openai-whisper = "^20231117"  # Local STT
piper-tts = {version = "^1.2.0", optional = true}  # Local TTS
pyttsx3 = "^2.90"  # Fallback TTS
pyaudio = "^0.2.14"  # Audio I/O
numpy = "^1.24.0"
soundfile = "^0.12.1"
webrtcvad = "^2.0.10"  # Voice activity detection
pvporcupine = {version = "^3.0.0", optional = true}  # Wake word detection

[tool.poetry.extras]
piper = ["piper-tts"]
wake = ["pvporcupine"]
full = ["piper-tts", "pvporcupine"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[tool.poetry.scripts]
zenus-voice = "zenus_voice.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
